Refactor test setup: Move get_init_db_payload and create_jwt_token to conftest.py for reuse; update tests for improved error handling in Guardian service responses.
